refactor a bit / tweak spacing
